% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.files_function.R
\name{input.files}
\alias{input.files}
\title{Reads files used in generating model input}
\usage{
input.files(ASE_file, legend_file, haplotypes_file)
}
\arguments{
\item{Allele-specific}{expression data; SNP information; haplotype information}
}
\description{
This function allows you to load all of your input files into memory
}
\examples{
input.files()
}
\keyword{files}
\keyword{input}
